fix(hook tests): fix intermittently failing hook tests
