--- Turner.Infrastructure.Crud.Tests/Fakes/FakeInjectable.cs
@@ -1,7 +1,0 @@
-﻿namespace Turner.Infrastructure.Crud.Tests.Fakes
-{
-    public class FakeInjectable
-    {
-        public int Value { get; set; }
-    }
-}
